new_demo/mts-teta-first-task/project/requirements.txt
pandas==
numpy==
matplotlib==
seaborn==
pyyaml==
tqdm==
mlflow==
scikit-learn==
lightgbm==
catboost==
sqlite==